ProjetPFE/backend/requirements.txt

motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
